+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   :warning: the job `update-datasets-reuses-metrics` disappears, you should unschedule it before installing this version
 - Consolidate CSW harvesting logic [#3378](https://github.com/opendatateam/udata/pull/3378)
   :warning: Config setting `HARVEST_ISO19139_XSL_URL` is renamed to `HARVEST_ISO19139_XSLT_URL`.
+- Ignore Dataset.accessService when processing DataService [#3399](https://github.com/opendatateam/udata/pull/3399)
 
 ## 10.8.2 (2025-07-31)